chore: Move horizontal methods to polars-ops
