+++ crates/polars-python/src/conversion/mod.rs
@@ -9,7 +9,6 @@ use std::path::PathBuf;
 #[cfg(feature = "object")]
 use polars::chunked_array::object::PolarsObjectSafe;
 use polars::frame::row::Row;
-use polars::frame::NullStrategy;
 #[cfg(feature = "avro")]
 use polars::io::avro::AvroCompression;
 #[cfg(feature = "cloud")]
